ligner polaroid

+++ lib/pages/story_image.dart
@@ -1,6 +1,7 @@
 
 import 'dart:async';
 import 'dart:io';
+import 'dart:math';
 import 'package:firebase_storage/firebase_storage.dart';
 import 'package:flutter/material.dart';
 import 'package:gap/gap.dart';
@@ -153,11 +154,15 @@ class _StoryScreenState extends State<StoryScreen> {
       //TODO den skal kun tjekke hvis der er noget fordi ellers kan man ikke slette
       String imagePath = photosData[0]['storagePath'];
 
-      try {
-        await storage.ref(imagePath).delete();
-      } catch (e) {
-        print(imagePath);
-        print('Failed to delete image: $e');
+      if (imagePath != "") {
+        try {
+          await storage.ref(imagePath).delete();
+        } catch (e) {
+          print(imagePath);
+          print('Failed to delete image: $e');
+        }
+      } else {
+        print('Image path is empty. No image to delete.');
       }
 
       docRef.delete();
@@ -222,6 +227,41 @@ Widget build(BuildContext context) {
           imagePath = findStoryImagePath(photosData[0]['story']);
         }
 
+        List<Color> backgroundColors = [
+        const Color(0xFFFFB5E8), 
+        const Color(0xFFFFE080), 
+        const Color(0xFFB2F7FF), 
+        const Color(0xFFC1FFC1), 
+        const Color(0xFFFFD8B1), 
+        const Color(0xFFD0A9FF), 
+        const Color(0xFFFFE5B4), 
+        const Color(0xFFB4EEB4), 
+        const Color(0xFFFFC0CB), 
+        const Color(0xFFB0E0E6), 
+        const Color(0xFFE6E6FA), 
+        const Color(0xFFFFD700), 
+        const Color(0xFF98FB98), 
+        const Color(0xFFFAFAD2), 
+        const Color(0xFFFFC3A0), 
+        const Color(0xFFFFFACD), 
+        const Color(0xFFA6E7FF), 
+        const Color(0xFF77DD77), 
+        const Color(0xFFFFE4E1), 
+        const Color(0xFFAB83A1), 
+        const Color(0xFF94E0FF), 
+        const Color(0xFFFFF8DC), 
+        const Color(0xFFAFEEEE), 
+        const Color(0xFFFFDEAD), 
+        const Color(0xFFFF9AA2), 
+        const Color(0xFFE599B1), 
+        const Color(0xFFDAA520), 
+        const Color(0xFFDA70D6), 
+        const Color(0xFFFF69B4), 
+        const Color(0xFFDB7093), 
+        const Color(0xFFCD5C5C)
+        ];
+      
+
         return Scaffold(
           resizeToAvoidBottomInset: false, 
           key: scaffoldKey,
@@ -232,36 +272,55 @@ Widget build(BuildContext context) {
             children: [
               const Gap(10),
               Text(
-                DateFormat("yyyy-MM-dd HH:mm").format(widget.timeTaken),
+                DateFormat("HH:mm, MMMM dd yyyy").format(widget.timeTaken),
                 style: TextStyle(
                   color: Theme.of(context).colorScheme.scrim,
                   fontSize: 26,
                   fontWeight: FontWeight.normal,
                 ),
               ),
-              SizedBox(
-                width: screenWidth - 150,
-                child: Image.file(
-                  File(widget.path),
-                  fit: BoxFit.cover,
+              Container(
+                color: backgroundColors[Random().nextInt(backgroundColors.length)],
+                padding: const EdgeInsets.all(25),
+                child: Column(
+                  children: [
+                    SizedBox(
+                    width: screenWidth - 150,
+                    height: screenWidth - 150,
+                      child: Image.file(
+                        File(widget.path),
+                        fit: BoxFit.cover,
+                      ),
+                    ),
+                    const Gap(25),
+                    description == '' ?
+                      OutlinedButton(
+                        onPressed: () {
+                          // Now call showAlert if photosData is not empty
+                          if (photosData.isNotEmpty) {
+                            showAlert(context, 'Change description');
+                          }
+                        },
+                        child: Text(
+                            'Add description',
+                            style: TextStyle(
+                              color: Theme.of(context).colorScheme.scrim,
+
+                            ),
+                          ),
+                      ) :
+                      Text(
+                        description,
+                        textAlign: TextAlign.center,
+                        style: TextStyle(
+                          fontFamily: 'CaveatBrush',
+                          color: Theme.of(context).colorScheme.scrim,
+
+                        ),
+                      ),
+                  ],
                 ),
-              ),
-              const Gap(10),
-               description == '' ?
-                OutlinedButton(
-                  onPressed: () {
-                    // Now call showAlert if photosData is not empty
-                    if (photosData.isNotEmpty) {
-                      showAlert(context, 'Change description');
-                    }
-                  },
-                  child: const Text('Add description')
-                ) :
-                SizedBox(
-                  width: screenWidth - 150,
-                  child: Text(description, textAlign: TextAlign.center,),
-                )
-              
+              )
             ],
           ),
         ),
